server/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==7.0.0
async-timeout==4.0.2
attrs==21.2.0
Automat==20.2.0
base58==2.1.1
beautifulsoup4==4.9.3
bitarray==1.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cytoolz==0.11.2
Django==2.2.24
django-cors-headers==3.10.0
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
extruct==0.13.0
frozenlist==1.2.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
h2==3.2.0
hexbytes==0.2.2
hpack==3.0.0
html-text==0.5.2
html5lib==1.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
ipfshttpclient==0.8.0a2
isodate==0.6.0
itemadapter==0.3.0
itemloaders==1.0.4
jmespath==0.10.0
jsonschema==3.2.0
jstyleson==0.0.2
lru-dict==1.1.7
lxml==4.6.3
mf2py==1.1.2
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
parsel==1.6.0
parsimonious==0.8.1
priority==1.3.0
promise==2.3
Protego==0.1.16
protobuf==3.19.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.12.0
PyDispatcher==2.0.5
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRdfa3==3.5.3
pyrsistent==0.18.0
pytz==2021.3
queuelib==1.6.1
rdflib==6.0.0
rdflib-jsonld==0.5.0
recipe-scrapers==13.3.5
requests==2.26.0
rlp==2.0.1
Rx==1.6.1
Scrapy==2.5.0
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
text-unidecode==1.3
toolz==0.11.2
Twisted==21.7.0
typing-extensions==3.10.0.0
urllib3==1.26.6
varint==1.0.2
w3lib==1.22.0
web3==5.25.0
webencodings==0.5.1
websockets==9.1
yarl==1.7.2
zope.interface==5.4.0
